Code refactoring.
* Rename files
* Rename APIs
