--- funsuite/src/main/scala/funsuite/FlakyFailure.scala
@@ -1,7 +1,0 @@
-package funsuite
-
-import scala.util.control.NoStackTrace
-
-class FlakyFailure(error: Throwable)
-    extends Exception("ignoring flaky test failure", error)
-    with NoStackTrace
