--- funsuite/src/main/scala/funsuite/Generic.scala
@@ -1,10 +1,0 @@
-package funsuite
-
-class GenericTest[T](
-    val name: String,
-    val body: () => T,
-    val tags: Set[Tag],
-    val location: Location
-) {
-  override def toString(): String = s"GenericTest($name)"
-}
