+++ funsuite/src/main/scala/funsuite/Printable.scala
@@ -1,5 +1,8 @@
 package funsuite
 
+/**
+  * Override this class to customize the default pretty-printer.
+  */
 trait Printable {
   def print(out: StringBuilder, indent: Int): Unit
 }
